requirements.txt
# ASGI / Server
fastapi==0.115.0
uvicorn[standard]==0.30.6

# Hugging Face / KoBERT
transformers==4.43.3
accelerate==0.33.0
sentencepiece==0.2.0
protobuf==3.20.3

# PyTorch (CUDA 11.8 build, GTX 1050 호환)
torch==2.7.1+cu118
torchvision==0.22.1+cu118
torchaudio==2.7.1+cu118

# Utils
numpy>=1.24
pydantic>=2.7
python-dotenv==1.0.1